requirements.txt
numpy>=1.26.4,<2.0.0
scipy==1.12.0
graspologic==3.4.1
future>=0.18.3
hyppo>=0.4.0
python-dotenv
tiktoken
pydantic
openai
pipmaster


aiohttp
configparser

# database packages
networkx

# Basic modules
numpy
pipmaster
pydantic
python-dotenv

setuptools
tenacity

# LLM packages
tiktoken

# Extra libraries are installed when needed using pipmaster
gensim